% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchnet-assent.R
\name{.assent_matrix}
\alias{.assent_matrix}
\title{Build the M x N confirmation-probability matrix}
\usage{
.assent_matrix(assent, M, N)
}
\arguments{
\item{assent}{A \code{saomnk_assent} object.}

\item{M, N}{Dimensions of the bipartite network.}
}
\value{
An \eqn{M \times N} matrix of probabilities in [0, 1].
}
\description{
Build the M x N confirmation-probability matrix
}
\keyword{internal}
