% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchnet-game.R
\name{.game_compute_scores}
\alias{.game_compute_scores}
\title{Compute scoreboard for all firms at the current state}
\usage{
.game_compute_scores(env, player_id)
}
\arguments{
\item{env}{SaomNkRSienaBiEnv object}

\item{player_id}{Integer, which firm is the player}
}
\value{
data.frame with per-firm scores
}
\description{
Compute scoreboard for all firms at the current state
}
\keyword{internal}
