% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchnet-game.R
\name{.game_reconstruct_snapshots}
\alias{.game_reconstruct_snapshots}
\title{Reconstruct bipartite matrix snapshots from the move log}
\usage{
.game_reconstruct_snapshots(game)
}
\arguments{
\item{game}{searchnet_game object}
}
\value{
list of M x N matrices, one per round (including initial)
}
\description{
Reconstruct bipartite matrix snapshots from the move log
}
\keyword{internal}
