% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nk-classic.R
\name{nk_fitness}
\alias{nk_fitness}
\title{Fitness of specified configurations}
\usage{
nk_fitness(landscape, config = NULL)
}
\arguments{
\item{landscape}{An \code{"nk_landscape"} object.}

\item{config}{Either an integer vector of configuration codes, or a binary
vector/matrix of configuration bits. If \code{NULL}, all fitness values
are returned.}
}
\value{
Numeric vector of fitness values.
}
\description{
Fitness of specified configurations
}
\examples{
nk <- nk_landscape(N = 6, K = 2, seed = 1)
nk_fitness(nk, c(1, 0, 1, 1, 0, 0))

}
