% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchnet-ergodicity.R
\name{print.searchnet_ergodicity}
\alias{print.searchnet_ergodicity}
\title{Print an Ergodicity Sweep}
\usage{
\method{print}{searchnet_ergodicity}(x, ...)
}
\arguments{
\item{x}{A \code{searchnet_ergodicity} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an Ergodicity Sweep
}
