% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchnet-replicator.R
\name{saomnk_plot_elo}
\alias{saomnk_plot_elo}
\title{Plot ELO Tournament Results}
\usage{
saomnk_plot_elo(
  tournament_result,
  colors = .policy_colors,
  title = "ELO Tournament Ratings"
)
}
\arguments{
\item{tournament_result}{List output from saomnk_elo_tournament}

\item{colors}{Named color vector}

\item{title}{Character plot title}
}
\value{
ggplot object
}
\description{
Bar chart of final ELO ratings with parity band.
}
