% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchnet-replicator.R
\name{saomnk_plot_simplex}
\alias{saomnk_plot_simplex}
\title{Plot Replicator Dynamics on Strategy Simplex}
\usage{
saomnk_plot_simplex(
  replicator_result,
  x_type = "RE",
  y_type = "RF",
  colors = .policy_colors,
  title = "Strategy Simplex"
)
}
\arguments{
\item{replicator_result}{Data frame from saomnk_replicator_dynamics}

\item{x_type}{Character name of policy type for x-axis}

\item{y_type}{Character name of policy type for y-axis}

\item{colors}{Named color vector}

\item{title}{Character plot title}
}
\value{
ggplot object
}
\description{
Creates a 2D De Finetti diagram (simplex projection) showing
evolutionary pressure and population trajectory.
}
