% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-unified.R
\name{searchnet-deprecated-plots}
\alias{searchnet-deprecated-plots}
\alias{saomnk_plot_exploration_exploitation}
\alias{saomnk_plot_exploration_exploitation_consistent}
\alias{saomnk_plot_exploration_exploitation_phase}
\alias{saomnk_plot_exploration_exploitation_improved}
\alias{saomnk_plot_exploration_exploitation_faceted}
\alias{saomnk_plot_exploration_exploitation_by_strategy}
\alias{saomnk_plot_strategy_exploration_exploitation}
\alias{saomnk_plot_exploration_exploitation_subsidies}
\alias{saomnk_search_rsiena_multiwave_plot}
\alias{saomnk_search_rsiena_multiwave_plot_K_4panel}
\alias{saomnk_search_rsiena_multiwave_plot_K_AA_strategy_summary}
\alias{saomnk_search_rsiena_multiwave_plot_K_AC_strategy_summary}
\alias{saomnk_search_rsiena_multiwave_plot_K_CA_strategy_summary}
\alias{saomnk_search_rsiena_multiwave_plot_K_CC_strategy_summary}
\alias{saomnk_search_rsiena_multiwave_plot_actor_utility_strategy_summary}
\alias{saomnk_search_rsiena_multiwave_plot_actor_utility_by_strategy}
\alias{saomnk_search_rsiena_multiwave_plot_actor_utility_density_by_strategy}
\alias{saomnk_search_rsiena_multiwave_plot_utility_ridge_density_by_strategy}
\alias{saomnk_plot_market_entry_survival_v0}
\alias{saomnk_plot_K_AC_NEW_shocks}
\title{Deprecated plot function names}
\usage{
saomnk_plot_exploration_exploitation(
  env,
  actor_ids = c(),
  thin_factor = 1,
  thin_pct = 1,
  smooth_method = "loess",
  show_points = TRUE,
  show_individuals = TRUE,
  show_group_means = TRUE,
  loess_span = 0.3,
  point_alpha_dimmer = 1,
  line_alpha = 0.5,
  group_line_size = 2,
  ylim = NULL,
  plot_return = TRUE,
  plot_save = FALSE,
  plot_file = "",
  plot_dir = NA
)

saomnk_plot_exploration_exploitation_consistent(
  env,
  actor_ids = c(),
  thin_factor = 1,
  thin_pct = 1,
  smooth_method = "loess",
  show_points = TRUE,
  show_individuals = FALSE,
  show_group_means = TRUE,
  loess_span = 0.3,
  point_alpha_dimmer = 1,
  line_alpha = 0.5,
  group_line_size = 2,
  se_ribbon = TRUE,
  ylim = c(0, 1),
  plot_return = TRUE,
  plot_save = FALSE,
  plot_file = "",
  plot_dir = NA
)

saomnk_plot_exploration_exploitation_phase(
  env,
  time_window = 50,
  show_trajectories = TRUE,
  plot_return = TRUE
)

saomnk_plot_exploration_exploitation_improved(
  env,
  actor_ids = c(),
  thin_factor = 1,
  show_points = FALSE,
  show_individuals = FALSE,
  show_group_means = TRUE,
  show_difference = TRUE,
  loess_span = 0.2,
  point_alpha = 0.2,
  line_alpha = 0.5,
  group_line_size = 2.5,
  se_ribbon = TRUE,
  ylim = c(0, 1),
  plot_return = TRUE,
  plot_save = FALSE,
  plot_file = "",
  plot_dir = NA
)

saomnk_plot_exploration_exploitation_faceted(
  env,
  metrics_df,
  loess_span = 0.3,
  show_points = FALSE
)

saomnk_plot_exploration_exploitation_by_strategy(
  env,
  actor_ids = c(),
  thin_factor = 1,
  show_points = TRUE,
  show_individuals = FALSE,
  show_group_means = TRUE,
  loess_span = 0.3,
  point_alpha = 0.3,
  line_alpha = 0.5,
  group_line_size = 2,
  se_ribbon = TRUE,
  ylim = c(0, 1),
  plot_return = TRUE,
  plot_save = FALSE,
  plot_file = "",
  plot_dir = NA
)

saomnk_plot_strategy_exploration_exploitation(
  env,
  result,
  show_points = TRUE,
  show_group_means = TRUE,
  group_line_size = 2
)

saomnk_plot_exploration_exploitation_subsidies(
  env,
  metrics_df = NULL,
  show_points = FALSE,
  show_se_ribbon = TRUE,
  loess_span = 0.3,
  shock_time = NULL
)

saomnk_search_rsiena_multiwave_plot(
  env,
  type = c(),
  rolling_window = 10,
  actor_ids = c(),
  component_ids = c(),
  wave_ids = c(),
  thin_factor = 1,
  thin_wave_factor = 1,
  smooth_method = "loess",
  show_utility_points = TRUE,
  show_strategy_means = TRUE,
  append_plot = FALSE,
  histogram_position = "identity",
  scale_utility = TRUE,
  return_plot = TRUE,
  plot_file = NA,
  plot_dir = NA,
  loess_span = 0.4
)

saomnk_search_rsiena_multiwave_plot_K_4panel(
  env,
  actor_ids = c(),
  component_ids = c(),
  wave_ids = c(),
  thin_factor = 1,
  thin_wave_factor = 1,
  smooth_method = "loess",
  show_utility_points = TRUE,
  return_plot = TRUE,
  plot_file = NA,
  plot_dir = NA
)

saomnk_search_rsiena_multiwave_plot_K_AA_strategy_summary(
  env,
  actor_ids = c(),
  wave_ids = c(),
  thin_factor = 1,
  thin_wave_factor = 1,
  smooth_method = "loess",
  show_utility_points = TRUE,
  show_legend = TRUE,
  show_title = TRUE,
  return_plot = TRUE,
  plot_file = NA,
  plot_dir = NA
)

saomnk_search_rsiena_multiwave_plot_K_AC_strategy_summary(
  env,
  actor_ids = c(),
  wave_ids = c(),
  thin_factor = 1,
  thin_wave_factor = 1,
  smooth_method = "loess",
  show_utility_points = TRUE,
  show_legend = TRUE,
  show_title = TRUE,
  return_plot = TRUE,
  plot_file = NA,
  plot_dir = NA
)

saomnk_search_rsiena_multiwave_plot_K_CA_strategy_summary(
  env,
  component_ids = c(),
  wave_ids = c(),
  thin_factor = 1,
  thin_wave_factor = 1,
  smooth_method = "loess",
  show_utility_points = TRUE,
  show_legend = TRUE,
  show_title = TRUE,
  return_plot = TRUE,
  plot_file = NA,
  plot_dir = NA
)

saomnk_search_rsiena_multiwave_plot_K_CC_strategy_summary(
  env,
  component_ids = c(),
  wave_ids = c(),
  thin_factor = 1,
  thin_wave_factor = 1,
  smooth_method = "loess",
  show_utility_points = TRUE,
  show_legend = TRUE,
  show_title = TRUE,
  return_plot = TRUE,
  plot_file = NA,
  plot_dir = NA
)

saomnk_search_rsiena_multiwave_plot_actor_utility_strategy_summary(
  env,
  actor_ids = c(),
  wave_ids = c(),
  thin_factor = 1,
  thin_wave_factor = 1,
  smooth_method = "loess",
  show_utility_points = TRUE,
  scale_utility = TRUE,
  return_plot = TRUE,
  plot_file = NA,
  plot_dir = NA,
  loess_span = 0.4
)

saomnk_search_rsiena_multiwave_plot_actor_utility_by_strategy(
  env,
  actor_ids = c(),
  thin_factor = 1,
  thin_wave_factor = 1,
  smooth_method = "loess",
  show_utility_points = TRUE,
  return_plot = TRUE,
  plot_file = NA,
  plot_dir = NA
)

saomnk_search_rsiena_multiwave_plot_actor_utility_density_by_strategy(
  env,
  thin_wave_factor = 1,
  return_plot = TRUE,
  plot_file = NA,
  plot_dir = NA
)

saomnk_search_rsiena_multiwave_plot_utility_ridge_density_by_strategy(
  env,
  actor_ids = c(),
  wave_ids = c(),
  thin_factor = 1,
  thin_wave_factor = 1,
  show_utility_points = TRUE,
  show_strategy_means = TRUE,
  scale_utility = TRUE,
  return_plot = TRUE,
  plot_file = NA,
  plot_dir = NA,
  plot_periods = 4
)

saomnk_plot_market_entry_survival_v0(
  env,
  n = 50,
  environ_params = NULL,
  structure_model = NULL,
  steps_per_actor = NULL,
  theta_shocks = NULL,
  conf_level = 0.95,
  verbose = FALSE
)

saomnk_plot_K_AC_NEW_shocks(env, verbose = FALSE)
}
\arguments{
\item{env}{A \code{SaomNkRSienaBiEnv} environment, passed on as the
replacement's first argument.}

\item{actor_ids, component_ids, wave_ids, thin_factor, thin_pct, thin_wave_factor, smooth_method, show_points, show_individuals, show_group_means, show_difference, loess_span, point_alpha, point_alpha_dimmer, line_alpha, group_line_size, se_ribbon, ylim, plot_return, plot_save, plot_file, plot_dir, time_window, show_trajectories, metrics_df, result, show_se_ribbon, shock_time, type, rolling_window, show_utility_points, show_strategy_means, append_plot, histogram_position, scale_utility, return_plot, show_legend, show_title, plot_periods, n, environ_params, structure_model, steps_per_actor, theta_shocks, conf_level, verbose}{Passed unchanged to the replacement; see its documentation.}
}
\value{
The replacement's return value.
}
\description{
These names are superseded under the package \link[=searchnet-naming]{naming policy}.
Each is a thin wrapper: it takes the same arguments as before, forwards
them to its replacement, and returns the same object. The first call in a
session emits a deprecation warning; later calls are silent.
}
\details{
\tabular{ll}{
   Deprecated \tab Replacement \cr
   \code{saomnk_plot_exploration_exploitation()} \tab \code{searchnet_plot_exploration(type = "trajectory")} \cr
   \code{saomnk_plot_exploration_exploitation_consistent()} \tab \code{searchnet_plot_exploration(type = "trajectory_density")} \cr
   \code{saomnk_plot_exploration_exploitation_phase()} \tab \code{searchnet_plot_exploration(type = "phase_space")} \cr
   \code{saomnk_plot_exploration_exploitation_improved()} \tab \code{searchnet_plot_exploration(type = "portfolio")} \cr
   \code{saomnk_plot_exploration_exploitation_by_strategy()} \tab \code{searchnet_plot_exploration(type = "strategy_groups")} \cr
   \code{saomnk_plot_exploration_exploitation_faceted()} \tab \code{searchnet_plot_exploration(type = "strategy_facets")} \cr
   \code{saomnk_plot_strategy_exploration_exploitation()} \tab \code{searchnet_plot_exploration(type = "strategy_result")} \cr
   \code{saomnk_plot_exploration_exploitation_subsidies()} \tab \code{searchnet_plot_exploration(type = "risk_subsidies")} \cr
   \code{saomnk_search_rsiena_multiwave_plot()} \tab \code{searchnet_plot_multiwave(what = "list")} \cr
   \code{saomnk_search_rsiena_multiwave_plot_K_4panel()} \tab \code{searchnet_plot_multiwave(what = "K_4panel")} \cr
   \code{saomnk_search_rsiena_multiwave_plot_K_AA_strategy_summary()} \tab \code{searchnet_plot_multiwave(what = "K_AA", by = "strategy")} \cr
   \code{saomnk_search_rsiena_multiwave_plot_K_AC_strategy_summary()} \tab \code{searchnet_plot_multiwave(what = "K_AC", by = "strategy")} \cr
   \code{saomnk_search_rsiena_multiwave_plot_K_CA_strategy_summary()} \tab \code{searchnet_plot_multiwave(what = "K_CA", by = "strategy")} \cr
   \code{saomnk_search_rsiena_multiwave_plot_K_CC_strategy_summary()} \tab \code{searchnet_plot_multiwave(what = "K_CC", by = "strategy")} \cr
   \code{saomnk_search_rsiena_multiwave_plot_actor_utility_strategy_summary()} \tab \code{searchnet_plot_multiwave(what = "utility", by = "strategy")} \cr
   \code{saomnk_search_rsiena_multiwave_plot_actor_utility_by_strategy()} \tab \code{searchnet_plot_multiwave(what = "utility", by = "actor")} \cr
   \code{saomnk_search_rsiena_multiwave_plot_actor_utility_density_by_strategy()} \tab \code{searchnet_plot_multiwave(what = "utility_density", by = "strategy")} \cr
   \code{saomnk_search_rsiena_multiwave_plot_utility_ridge_density_by_strategy()} \tab \code{searchnet_plot_multiwave(what = "utility_ridge", by = "strategy")} \cr
   \code{saomnk_plot_market_entry_survival_v0()} \tab \code{searchnet_plot_cumulative_entry()} \cr
   \code{saomnk_plot_K_AC_NEW_shocks()} \tab \code{searchnet_plot_new_component_shocks()} \cr
}
}
\seealso{
\code{\link[=searchnet_plot_exploration]{searchnet_plot_exploration()}}, \code{\link[=searchnet_plot_multiwave]{searchnet_plot_multiwave()}},
\code{\link[=searchnet_plot_cumulative_entry]{searchnet_plot_cumulative_entry()}},
\code{\link[=searchnet_plot_new_component_shocks]{searchnet_plot_new_component_shocks()}}, \link{searchnet-naming}.
}
\keyword{internal}
