% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchnet-teaching.R
\name{.create_ai_opponents}
\alias{.create_ai_opponents}
\title{Create AI firm decision logic}
\usage{
.create_ai_opponents(env, model, n_ai, actor_ids)
}
\arguments{
\item{env}{SaoMNK environment}

\item{model}{saomnk_model object}

\item{n_ai}{Number of AI actors}

\item{actor_ids}{Integer vector of AI actor IDs}
}
\value{
List with AI configuration
}
\description{
Create AI firm decision logic
}
\keyword{internal}
