% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchnet-replicator.R
\docType{data}
\name{searchnet-replicator}
\alias{searchnet-replicator}
\alias{.policy_colors}
\title{Replicator Dynamics and Evolutionary Game Theory}
\format{
An object of class \code{character} of length 4.
}
\usage{
.policy_colors
}
\description{
Functions for running replicator dynamics, ELO tournaments,
and computing evolutionarily stable strategies (ESS) for policy type
populations in the SaoMNK framework.
}
\keyword{datasets}
